requirements.txt
py-clob-client>=0.28.0
python-dotenv>=1.0.0
requests>=2.31.0
